consolidate stats
